api files refactoring
